feat: Add automerge workflow (default on)
Adds an automerge GH workflow to the project. This workflow depends
on the automerge setting in Github to be on, and the upgrade workflow
to have 'acceptance' as its base branch.

NB: To prevent unwanted automatic merges, branch protection should be
on for acceptance, and to require succesful builds. Otherwise,
unsuccesful dependency updates may merge to acceptance.